Move migrators in their own modules.
